refactor(shell): isolate workspace command analysis
